dclxvi-20130329: Performance fix!
Got your attention? Fix a bug/typo in commit
2b4b2750cf3702c2b8583fb342597ac339b9bf20
which was causing the library to use the C-implementation, not
the assembly one.

With this change, we see the order of magnitude improvement in
pairing performance that we were expecting:
BenchmarkPairGo-6             50          34319034 ns/op
BenchmarkPairCGO-6          1000           1565629 ns/op <-- 20x better!
BenchmarkPairC-6            1000           1510599 ns/op