dclxvi-20130329: Ability to build library with pure-C implementation.
As is, the optimized assembly seems to work only on linux/amd64.
On OS X for example, clang isn't happy with the instructions
(giving an error:
  gcc -std=c99 -O3 -fomit-frame-pointer -c -o fp2e_double2.o fp2e_double2.s
  clang -cc1as: fatal error: error in backend: 32-bit absolute addressing is not supported in 64-bit mode architectures
)

With this commit, can revert to the pure-C implementation by adding the
USE_C=true flag before make.

Benchmarks suggest that the pure-C implementation is about 3x
faster than the pure-Go implementation, but 6x slower than optimized
assembly.