android: Use bendroid instead of gomobile.
This commit switches to using https://godoc.org/v.io/x/devtools/bendroid
instead of gomobile for running benchmarks on Android. The main
motivation is that doing so allows the exact same test and benchmark
code that is used for desktops/Pis/laptops to be used for Android.
The code duplicated in the android/ subdirectory goes away.